pyproject.toml
[tool.poetry]
name = "discord-url-bypass-bot"
version = "0.1.0"
description = "A Discord bot that unlocks protected URLs using headless browser automation."
authors = ["YourName <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
discord.py = "^2.3.2"
undetected-chromedriver = "^3.5.5"
python-dotenv = "^1.0.1"
selenium = "^4.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"